Fix broken audio in videos for Tokyo Necro
Cleanly fixes the audio not playing for in-game videos via the GST_PLUGIN_FEATURE_RANK environment variable. 

To this date, the developers still hasn't fixed this problem and a popular reported workaround for this was to move Proton's media converter shared library file. 

Setting this variable eliminates the need for that.