Grim Dawn: remove deprecated mod
I have investigated and it seems that "Grim Internals" is a mod now deprecated by an actual update to the game. It does not work anymore on new versions of the game. For people that want to run it on older versions, they can just run the executable manually.

Source of the information: https://forums.crateentertainment.com/t/tool-grim-internals/38773/7429