patch_libcuda: Fix patched library name and use a temporary directory instead of .cache (#337)
* patch_libcuda: Ensure the patched library name is the same as the library we found.

This should ensure that this library is loaded correctly as it's path is added to the beginning of LD_LIBRARY_PATH.

* patch_libcuda: Create a temporary directory and place the patched library there.

This avoids using the cache, as we already clobber the library that would be placed there.

* patch_libcuda: We don't need to worry about the cache directory existing anymore.

* patch_libcuda: Set LD_PRELOAD as well.

This is useful in the case that the LD_LIBRARY_PATH workaround does not work. LD_PRELOAD works in an environment where EAC isn't running or is set to run as a null client.

* patch_libcuda: Slightly refactor how environment variables are set.

This is cleaner and makes reading this easier.