Add fix for Not For Broadcast (#437)
This game is heavily reliant on video playback, so much so that it doesn't work with default Proton. Proton-GE is needed.

There are some beta branches of the game that aim to fix video playback that some users experience. I've tried only one of them, but it didn't help me.

Then I remembered about the GST media thing and I tried the normal version (without beta branches selected) and in comparison to winedmo backend it seems all video/game freezing and crashing is fixed.

https://store.steampowered.com/app/1147550/Not_For_Broadcast/