upscaler: make more robust against network errors and add more logging
Unlike the other upscalers, the FSR4 dll does not have a checksum to
validate it. This may result in a faulty file being copied to the prefix
without any error. To help reduce the risk of that happening, if an
exception occurs while downloading a file, remove the file from the
cache directory. For the files that have an md5 checksum, validate the
downloaded file after completion.

To make sure the dlls installed in the prefix are correct, store the md5
hash of file in the version file. When running the post-download check,
check the files against the stored hash of the existing version to validate
that they have not been corrupted somehow.