Update fix for Angelic Chaos: RE-BOOT! (#475)
A change in Valve's wine media playback code in Proton 10 caused in-game media playback to no longer work and crash the game instead. To restore media playback and not crash, refer to the built-in hack for ATRI -My Dear Moments-, use the gstreamer-based backend, and configure the game's movie rendering.