Update fix for Full Metal Daemon Muramasa
Attempting to play the videos from the game's Extra menu with the current fix results in the videos being skipped. At the time, this was due to the fix disabling the wmvdecod DLL, which was needed to get in-game media to render otherwise placeholders would be displayed. Since then both GE-Proton and Proton (bleeding-edge) has improved a bit. On the other hand, without applying the fix, the videos would play but without any audio. To ensure that both the videos listed in the game's Extra menu and in-game media renders properly, use the gstreamer-based backend. Tested on the Steam Deck running with SteamOS (3.7.19) and using GE-Proton10-26.